Add check for existed id feature

+++ script.js
@@ -43,7 +43,7 @@ function renderTable() {
         <td>${contact.details}</td>
         <td>
           <button onclick="deleteContact(${contact.id})" class="btn back-red text-moon"> <i class='fa fa-trash'></i></button>
-          <button class="btn back-green text-moon"> <i class='fa fa-edit'></i></button>
+          <button onclick="updateStart(${contact.id})" class="btn back-green text-moon"> <i class='fa fa-edit'></i></button>
         </td>`
     );
     tbody.append(tr);
@@ -75,12 +75,23 @@ $("#contact_form").on("submit", (event) => {
   for (item of newContact.entries()) {
     new_contact[item[0]] = item[1];
   }
-  // Push it into cotact list and updated list into localstorage
-  contactList.push(new_contact);
-  localStorage.setItem("contactList", JSON.stringify(contactList));
-  // Update view of html
-  hideForm();
-  renderTable();
+  // Check if the id existed
+  if (contactList.find((item) => item.id == new_contact.id)) {
+    Swal.fire({
+      title: "Id Existed",
+      text: "Your chosen id is already existed choose another id!",
+      icon: "warning",
+      confirmButtonColor: "#14C2A3",
+      confirmButtonText: "Ok",
+    });
+  } else {
+    // Push it into cotact list and updated list into localstorage
+    contactList.push(new_contact);
+    localStorage.setItem("contactList", JSON.stringify(contactList));
+    // Update view of html
+    hideForm();
+    renderTable();
+  }
 });
 
 function deleteContact(id) {
@@ -104,4 +115,8 @@ function deleteContact(id) {
       });
     }
   });
+}
+
+function updateStart(id) {
+  console.log(id);
 }
